+++ CHANGELOG.md
@@ -13,7 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2019-10-27
 ### Added
+- Add getLocalStorage command
+- Add setLocalStorage command
+- Add removeLocalStorage command
 - Add acceptance tests
+
+### Changed
+- Improve documentation
 
 ## [1.0.0] - 2019-10-26
 ### Added